Fix content-type header for Playlist#add_tracks.
When adding a track to a playlist I was getting the following warning

> warning: Overriding "Content-Type" header "application/json" with "application/x-www-form-urlencoded" due to payload

It looks like `RestClient` assumed that an empty hash should have been form encoded rather than JSON encoded and showed this warning. Explicitly turning the empty params hash `to_json` quietened down the warning and meant that the API calls were being made with the correct Content-Type.